lint

+++ components/raftstore/src/store/peer.rs
@@ -5152,7 +5152,7 @@ where
         };
         let state = if !self.replication_sync {
             if self.dr_auto_sync_state != DrAutoSyncState::Async {
-                let res = self.check_group_commit_consistent(5 as u64);
+                let res = self.check_group_commit_consistent(5_u64);
                 if Some(true) != res {
                     let mut buffer: SmallVec<[(u64, u64, u64); 5]> = SmallVec::new();
                     if self.get_store().applied_term() >= self.term() {
